requirements.txt
requests~=2.31.0
python-dotenv~=1.0.1
exchangelib~=5.1.0